Homeworks/BookRecommender/src/bg/sofia/uni/fmi/mjt/goodreads/recommender/similaritycalculator/composite/CompositeSimilarityCalculator.java
package bg.sofia.uni.fmi.mjt.goodreads.recommender.similaritycalculator.composite;

import bg.sofia.uni.fmi.mjt.goodreads.book.Book;
import bg.sofia.uni.fmi.mjt.goodreads.recommender.similaritycalculator.SimilarityCalculator;

import java.util.Map;

public class CompositeSimilarityCalculator implements SimilarityCalculator {
    private final Map<SimilarityCalculator, Double> similarityCalculators;

    public CompositeSimilarityCalculator(Map<SimilarityCalculator, Double> similarityCalculatorMap) {
        similarityCalculators = similarityCalculatorMap;
    }

    @Override
    public double calculateSimilarity(Book first, Book second) {
        validateBook(first);
        validateBook(second);

        return similarityCalculators
            .entrySet()
            .stream()
            .mapToDouble(e -> e.getValue() * e.getKey().calculateSimilarity(first, second))
            .sum();
    }

    private void validateBook(Book book) {
        if (book == null) {
            throw new IllegalArgumentException("Book cannot be null!");
        }
    }
}